PAYARA-2143-Payara-Micro-maven-plugin-add-skip-property

+++ payara-micro-maven-plugin/src/main/java/fish/payara/maven/plugins/micro/StopMojo.java
@@ -69,6 +69,11 @@ public class StopMojo extends BasePayaraMojo {
 
     @Override
     public void execute() throws MojoExecutionException, MojoFailureException {
+        if (skip) {
+            getLog().info("Stop mojo execution is skipped");
+            return;
+        }
+
         if (processId != null) {
             killProcess(processId);
         }
